[go] appencryption integration tests improvements
Integration tests, namely those dependant on testcontainers, have
been relocated to a new module, eliminating many test-related dependencies.

A new workspace has been defined to simplify working with nested modules (go.work).
