+++ go/appencryption/go.work
@@ -1,0 +1,7 @@
+go 1.19
+
+use (
+	.
+	./cmd/example
+	./integrationtest
+)
