+++ go/appencryption/integrationtest/integration_benchmark_test.go
@@ -1,4 +1,4 @@
-package appencryption_test
+package integration_test
 
 import (
 	"context"
